refactor(git): Extract error handling to separate file
